+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -4,6 +4,7 @@
 -   Fix styling of the advanced filter operator selection. #7005
 -   Remove the use of Dashicons and replace with @wordpress/icons or gridicons #7020
 -   Add tree shaking support to this package. #7034
+-   Deprecate the Gravatar component. #7716
 -   Remove useFilters from the package. #7117
 -   Deprecate SegmentedSelection, it will be removed in the next major release. #7118
 -   Deprecate the Count component, with plan to remove in next major version. #7115